+++ src/App/Http/Controllers/Api/UserController.php
@@ -48,15 +48,16 @@ class UserController extends Controller
                     ->column(['prop' => 'name',       'label'=> '用户名', 'minWidth'=> '120'])
                     ->column(['prop' => 'email',      'label'=> '邮箱',   'minWidth'=> '180'])
                     ->column(['prop' => 'mobile',     'label'=> '手机',   'minWidth'=> '180'])
-                    ->column(['prop' => 'status',     'label'=> '状态',   'width'=> '90',      'type' => 'status', 'config' => $this->userModel->status])
+                    ->column(['prop' => 'status',     'label'=> '状态',   'width'=> '90',      'type' => 'status'])
                     ->column(['prop' => 'rightButton','label'=> '操作',   'minWidth'=> '220',  'type' => 'btn'])
                     ->topButton(['buttonType'=>'add',        'apiUrl'=> route('api.admin.user.user.add'),'title'=>'新增用户','icon'=>'fa fa-plus'])                         // 添加新增按钮
-                    ->topButton(['buttonType'=>'resume',     'apiUrl'=> route('api.admin.user.user.status')])                         // 添加启用按钮
-                    ->topButton(['buttonType'=>'forbid',     'apiUrl'=> route('api.admin.user.user.status')])                         // 添加禁用按钮
-                    ->topButton(['buttonType'=>'delete',     'apiUrl'=> route('api.admin.user.user.delete')])                         // 添加删除按钮
+                    ->topButton(['buttonType'=>'open',    'apiUrl'=> route('api.admin.user.user.status'), 'data'=>'open'])                         // 添加启用按钮
+                    ->topButton(['buttonType'=>'close',    'apiUrl'=> route('api.admin.user.user.status'), 'data'=>'close'])                         // 添加禁用按钮
+                    ->topButton(['buttonType'=>'delete',    'apiUrl'=> route('api.admin.user.user.delete'), 'data'=>'delete'])
                     ->rightButton(['buttonType'=>'edit',     'apiUrl'=> route('api.admin.user.user.edit')])                         // 添加编辑按钮
-                    ->rightButton(['buttonType'=>'forbid',   'apiUrl'=> route('api.admin.user.user.status')])                       // 添加禁用/启用按钮
-                    ->rightButton(['buttonType'=>'delete',   'apiUrl'=> route('api.admin.user.user.delete')])                       // 添加删除按钮
+                    ->rightButton(['buttonType'=>'open',  'apiUrl'=> route('api.admin.user.user.status'), 'show'=>['close'], 'data'=>'open' ])                       // 添加禁用/启用按钮
+                    ->rightButton(['buttonType'=>'close',  'apiUrl'=> route('api.admin.user.user.status'), 'show'=>['open'], 'data'=>'close' ])                       // 添加禁用/启用按钮
+                    ->rightButton(['buttonType'=>'delete',  'apiUrl'=> route('api.admin.user.user.delete'), 'data'=>'delete'])
                     ->pagination(['total'=>$data['total'], 'pageSize'=>$data['pageSize'], 'pageSizes'=>$pageSizes])//分页设置
                     ->searchTitle('请输入搜索内容')
                     ->searchSelect(['id'=>'ID','name'=>'用户名','email'=>'邮箱','mobile'=>'手机'])
@@ -67,19 +68,29 @@ class UserController extends Controller
     {
         $input = $request->all();
         foreach ($input as $id => $value) {
-            $this->userModel->where('id', '=', $id)->update(['status' => $value]);
+            if ($value == 'close' || $value == 'open') {
+                $this->userModel->where('id', '=', $id)->update(['status' => $value]);
+            }
         }
-        $message = [
-                                        'message'   => '用户状态更改成功!',
-                    'type'      => 'success',
-                                    ];
+        if ($value == 'close') {
+            $message = [
+                'message'   => '用户状态关闭成功!',
+                'type'      => 'success',
+            ];
+        } elseif ($value == 'open') {
+            $message = [
+                'message'   => '用户状态开启成功!',
+                'type'      => 'success',
+            ];
+        }
         return resolve('builderHtml')->message($message)->response();
     }
     public function delete(Request $request)
     {
-        $input = $request->all();
-        foreach ($input as $id => $value) {
-            $response = $this->userModel->find($id)->forceDelete();
+        foreach ($request->all() as $id => $value) {
+            if ($value == 'delete') {
+                $response = $this->userModel->find($id)->forceDelete();
+            }
         }
         $message = [
                     'message'   => '后台用户删除成功!',
